Skeleton code for NN; reformatted P1; Adds data generation code (todo finish)

+++ Project2-Stream-Compaction/src/testing_helpers.hpp
@@ -7,8 +7,8 @@
 #include <ctime>
 
 template<typename T>
-int cmpArrays(int n, T *a, T *b) {
-    for (int i = 0; i < n; i++) {
+int cmpArrays(unsigned long int n, T *a, T *b) {
+    for (unsigned long int i = 0; i < n; i++) {
         if (a[i] != b[i]) {
             printf("    a[%d] = %d, b[%d] = %d\n", i, a[i], i, b[i]);
             return 1;
@@ -22,13 +22,13 @@ void printDesc(const char *desc) {
 }
 
 template<typename T>
-void printCmpResult(int n, T *a, T *b) {
+void printCmpResult(unsigned long int n, T *a, T *b) {
     printf("    %s \n",
             cmpArrays(n, a, b) ? "FAIL VALUE" : "passed");
 }
 
 template<typename T>
-void printCmpLenResult(int n, int expN, T *a, T *b) {
+void printCmpLenResult(unsigned long int n, unsigned long int expN, T *a, T *b) {
     if (n != expN) {
         printf("    expected %d elements, got %d\n", expN, n);
     }
@@ -38,29 +38,29 @@ void printCmpLenResult(int n, int expN, T *a, T *b) {
 }
 
 
-void zeroArray(int n, int *a) {
-    for (int i = 0; i < n; i++) {
+void zeroArray(unsigned long int n, int *a) {
+    for (unsigned long int i = 0; i < n; i++) {
         a[i] = 0;
     }
 }
 
-void onesArray(int n, int *a) {
-	for (int i = 0; i < n; i++) {
+void onesArray(unsigned long int n, int *a) {
+	for (unsigned long int i = 0; i < n; i++) {
 		a[i] = 1;
 	}
 }
 
-void genArray(int n, int *a, int maxval) {
+void genArray(unsigned long int n, int *a, int maxval) {
     srand(time(nullptr));
 
-    for (int i = 0; i < n; i++) {
+    for (unsigned long int i = 0; i < n; i++) {
         a[i] = rand() % maxval;
     }
 }
 
-void printArray(int n, int *a, bool abridged = false) {
+void printArray(unsigned long int n, int *a, bool abridged = false) {
     printf("    [ ");
-    for (int i = 0; i < n; i++) {
+    for (unsigned long int i = 0; i < n; i++) {
         if (abridged && i + 2 == 15 && n > 16) {
             i = n - 2;
             printf("... ");
