+++ Project2-Stream-Compaction/stream_compaction/cpu.cu
@@ -17,10 +17,10 @@ namespace StreamCompaction {
          * For performance analysis, this is supposed to be a simple for loop.
          * (Optional) For better understanding before starting moving to GPU, you can simulate your GPU scan in this function first.
          */
-        void scan(int n, int *odata, const int *idata, const bool time/* = true*/) {
+        void scan(unsigned long int n, int *odata, const int *idata, const bool time/* = true*/) {
 	        (time == true)?timer().startCpuTimer():0;
-			int running_total = 0;
-			for (int i = 0; i < n-1; i++) { // n-1 because we want exclusive not inclusive so last element isn't added
+			unsigned long int running_total = 0; // assumes only positive values
+			for (unsigned long int i = 0; i < n-1; i++) { // n-1 because we want exclusive not inclusive so last element isn't added
 				running_total += idata[i]; // n-1 adds for n size 
 				odata[i+1] = running_total;
 			}
@@ -33,11 +33,11 @@ namespace StreamCompaction {
          *
          * @returns the number of elements remaining after compaction.
          */
-        int compactWithoutScan(int n, int *odata, const int *idata) {
+		unsigned long int compactWithoutScan(unsigned long int n, int *odata, const int *idata) {
 	        timer().startCpuTimer();
-			int count = 0;
-			int odata_pos = 0;
-			for (int i = 0; i < n; i++)
+			unsigned long int count = 0;
+			unsigned long int odata_pos = 0;
+			for (unsigned long int i = 0; i < n; i++)
 				if (idata[i] != 0) {
 					count++;
 					odata[++odata_pos] = idata[i];
@@ -51,12 +51,12 @@ namespace StreamCompaction {
          *
          * @returns the number of elements remaining after compaction.
          */
-        int compactWithScan(int n, int *odata, const int *idata) {
+		unsigned long int compactWithScan(unsigned long int n, int *odata, const int *idata) {
 	        timer().startCpuTimer();
 	        // Step 1, mark each cell with 1/0 if it has a element or not (super easy to parallelize)
 			int* scan_data = new int[n]();
 			int* mask = new int[n]();
-			for (int i = 0; i < n; i++) {
+			for (unsigned long int i = 0; i < n; i++) {
 				if (idata[i])
 					mask[i] = 1;
 				else
@@ -65,7 +65,7 @@ namespace StreamCompaction {
 			// scan the mask array (can be done in parallel by using a balanced binary tree)
 			scan(n, scan_data, mask, false);
 			// Scatter array (go to each position and copy the value) (super easy to parallelize)
-			for (int i = 0; i < n - 1; i++) {
+			for (unsigned long int i = 0; i < n - 1; i++) {
 				if (idata[i])
 					odata[scan_data[i] + 1] = idata[i]; 
 			}
