+++ Project2-Stream-Compaction/stream_compaction/cpu.h
@@ -6,10 +6,10 @@ namespace StreamCompaction {
     namespace CPU {
 	    StreamCompaction::Common::PerformanceTimer& timer();
 
-        void scan(int n, int *odata, const int *idata, const bool time = true);
+        void scan(unsigned long int n, int *odata, const int *idata, const bool time = true);
 
-        int compactWithoutScan(int n, int *odata, const int *idata);
+		unsigned long int compactWithoutScan(unsigned long int n, int *odata, const int *idata);
 
-        int compactWithScan(int n, int *odata, const int *idata);
+		unsigned long int compactWithScan(unsigned long int n, int *odata, const int *idata);
     }
 }
